feat: replace NLTK with spaCy for sentence splitting
NLTK's sent_tokenize required a separate punkt_tab data download and
LookupError handling. spaCy's rule-based sentencizer ships with the
core package, removing that friction while providing comparable
sentence boundary detection.
