+++ docs/extraction/chunkers.md
@@ -126,19 +126,16 @@ Chunk 1: "Sentence three. Sentence four. Sentence five."
 - Articles, reports, documentation
 - Content you might show to users
 
-#### NLTK-Enhanced Sentence Splitting
+#### spaCy-Enhanced Sentence Splitting
 
-The semantic chunker can optionally use `nltk.tokenize.sent_tokenize` for more accurate sentence boundary detection. This handles abbreviations (e.g., "Dr. Smith"), decimal numbers (e.g., "3.14"), and URLs better than the default regex splitter.
+The semantic chunker can optionally use spaCy's `sentencizer` component for more accurate sentence boundary detection. The sentencizer is a rule-based component that ships with spaCy core — no separate model download needed.
 
 ```bash
 # Install the optional NLP extra
 pip install khora[nlp]
-
-# Download the required tokenizer data
-python -m nltk.downloader punkt_tab
 ```
 
-When NLTK is installed and the `punkt_tab` data is available, the semantic chunker uses it automatically — no code changes needed. If NLTK is not installed or the data is missing, the chunker falls back to its regex-based splitter transparently. The `_HAS_NLTK` flag in the chunker module controls this behavior.
+When spaCy is installed, the semantic chunker uses it automatically — no code changes needed. If spaCy is not installed, the chunker falls back to its regex-based splitter transparently. The `_HAS_SPACY` flag in the chunker module controls this behavior.
 
 ### Recursive Chunker
