+++ pyproject.toml
@@ -85,7 +85,7 @@ accel = [
     "rapidfuzz>=3.0.0",
 ]
 # NLP tools
-nlp = ["nltk>=3.9.0"]
+nlp = ["spacy>=3.8.0"]
 # Rust-accelerated operations (requires Rust toolchain to build from source)
 rust = [
     "khora-accel>=0.1.0",
@@ -193,5 +193,5 @@ dev = [
     "isort>=7.0.0",
     "ty>=0.0.15",
     "rapidfuzz>=3.0.0",  # For ty type checking of optional accel module
-    "nltk>=3.9.0",
+    "spacy>=3.8.0",
 ]
